Creada una interfaz IDraggable
Creada una interfaz con el metodo dragGO, eliminado el script de Draggable para optimizacion
