+++ Assets/Prefabs/Block.prefab
@@ -13,7 +13,6 @@ GameObject:
   - component: {fileID: 7544763875793729748}
   - component: {fileID: -7110608547483943812}
   - component: {fileID: -5511280168191358873}
-  - component: {fileID: -8279106895619994170}
   - component: {fileID: -72412387645844089}
   m_Layer: 7
   m_Name: Block
@@ -135,24 +134,6 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 112
   m_CollisionDetection: 1
---- !u!114 &-8279106895619994170
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8402551008408309344}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 798d34e3b4151194490e786a6a9d392f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  bc: {fileID: -7110608547483943812}
-  rb: {fileID: -5511280168191358873}
-  whatToDetect:
-    serializedVersion: 2
-    m_Bits: 128
-  isDoll: 0
 --- !u!114 &-72412387645844089
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -166,3 +147,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   rb: {fileID: -5511280168191358873}
+  blockLayer:
+    serializedVersion: 2
+    m_Bits: 128
