+++ Assets/Scripts/Bloques/Doll.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 30cdf78b5bee51a49a42711ae9f3047c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
