+++ Assets/Scripts/Fase construcción/IDraggable.cs
@@ -1,0 +1,4 @@
+public interface IDraggable
+{
+    void dragGO();
+}
