+++ Assets/Scripts/Fase construcción/IDraggable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 798d34e3b4151194490e786a6a9d392f
+guid: 9a720e600257e8f4eac880072b5510d0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
